Move terminal bits out of simpleCompletionModel
Part of #238986
